pyproject.toml
[build-system]
build-backend = "flit_core.buildapi"
requires = ["flit_core >=3.8.0,<4"]

[project]
name = "mediaify"
version = "3.0.0"
description = "Media encoding made simple"
license = {file = "LICENSE"}
readme = "README.md"
authors = [
    {name = "Ben Brady", email = "benbradybusiness@gmail.com"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Multimedia",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Multimedia :: Graphics :: Graphics Conversion",
    "Topic :: Multimedia :: Video",
    "Topic :: Multimedia :: Video :: Conversion",
    "Typing :: Typed",
]
requires-python = ">=3.9"
dependencies = [
    "python-ffmpeg~=2.0.4",
    "MediaInfo~=0.0.9",
    "pillow~=9.5.0",
    "python_magic~=0.4.27",
    "python-magic-bin; sys_platform == 'win32'",
    "numexpr",
]
[project.urls]
Source = "https://github.com/Ben-Brady/mediaify"
Documentation = "https://mediaify.readthedocs.io/"


[project.optional-dependencies]
test = [
    "mypy ==0.982",
    "pytest >=7.1.3,<8.0.0",
    "black == 23.1.0",
    # "coverage[toml] >= 6.5.0,< 8.0",
    # "ruff ==0.0.138",
]
docs = [
    "libmagic",
    # "coverage[toml] >= 6.5.0,< 8.0",
    # "ruff ==0.0.138",
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]

[tool.mypy]
warn_return_any = true
warn_unused_configs = true
strict = true
files = ["mediaify"]
